+++ src/library/UnityHacks.h
@@ -23,11 +23,37 @@
 #include <sys/types.h>
 #include <cstddef>
 #include <cstdint>
+#include <vector>
+#include <map>
+#include <string>
 
 namespace libtas {
 
 namespace UnityHacks
 {
+
+// From ImPlot
+struct ScrollingBuffer {
+    size_t MaxSize;
+    int Offset;
+    std::vector<float> DataX;
+    std::vector<float> DataY;
+    std::string name;
+
+    ScrollingBuffer(int max_size);
+    void AddPoint(int x, int y);
+    void Erase();
+};
+
+struct ScrollingBuffers {
+    std::map<int, ScrollingBuffer> Buffers;
+    
+    ScrollingBuffers();
+    void AddPoint(float x, float y, int tid);
+};
+
+const ScrollingBuffers& getJobData();
+    
 void setUnity();
 
 bool isUnity();